Enum for keywords, simplify symbol types, Expect() Check for keywords directly instead of indirectly via their symbol types In particular, introduce named constand for "no symbol" instead of using hardcoded magic number Therefore, change SkipTo() to work with symbol lists instead of symbol type lists Give all keywords the symbol type kSYM_Keyword instead of their own symbol type so that "symbol type" means what it should mean, to be a "type" Don't misuse symbol types for nesting stack entry types Use Expect() to check for symbols that must come in the input

+++ Compiler/test2/cc_parser_test_1.cpp
@@ -906,7 +906,7 @@ TEST_F(Compile1, FixupMismatch) {
     // Sanity check for import fixups: The corresponding Bytecode must
     // point into the imports[] array, and the corresponding slot must
     // contain a non-empty string.
-    for (size_t fixup_idx = 0; fixup_idx < scrip.numfixups; fixup_idx++)
+    for (size_t fixup_idx = 0; fixup_idx < static_cast<size_t>(scrip.numfixups); fixup_idx++)
     {
         if (FIXUP_IMPORT != scrip.fixuptypes[fixup_idx])
             continue;
